refactoring
- c to cpp
- code units delegated to the corresponding files with headers
