+++ .idea/.name
@@ -1,0 +1,1 @@
+simple_text_editor
