--- cmake-build-debug/.cmake/api/v1/reply/codemodel-v2-03f7bf020658a9c73a23.json
@@ -1,60 +1,0 @@
-{
-	"configurations" : 
-	[
-		{
-			"directories" : 
-			[
-				{
-					"build" : ".",
-					"jsonFile" : "directory-.-Debug-f5ebdc15457944623624.json",
-					"minimumCMakeVersion" : 
-					{
-						"string" : "3.28"
-					},
-					"projectIndex" : 0,
-					"source" : ".",
-					"targetIndexes" : 
-					[
-						0
-					]
-				}
-			],
-			"name" : "Debug",
-			"projects" : 
-			[
-				{
-					"directoryIndexes" : 
-					[
-						0
-					],
-					"name" : "simple_text_editor",
-					"targetIndexes" : 
-					[
-						0
-					]
-				}
-			],
-			"targets" : 
-			[
-				{
-					"directoryIndex" : 0,
-					"id" : "simple_text_editor::@6890427a1f51a3e7e1df",
-					"jsonFile" : "target-simple_text_editor-Debug-ca2e4020dfded9495b25.json",
-					"name" : "simple_text_editor",
-					"projectIndex" : 0
-				}
-			]
-		}
-	],
-	"kind" : "codemodel",
-	"paths" : 
-	{
-		"build" : "/Users/owner/Storage/KSE/IT Disciplines/PP/simple-text-editor/cmake-build-debug",
-		"source" : "/Users/owner/Storage/KSE/IT Disciplines/PP/simple-text-editor"
-	},
-	"version" : 
-	{
-		"major" : 2,
-		"minor" : 6
-	}
-}
