--- cmake-build-debug/.cmake/api/v1/reply/toolchains-v1-b5b70a09b5cb5c487388.json
@@ -1,44 +1,0 @@
-{
-	"kind" : "toolchains",
-	"toolchains" : 
-	[
-		{
-			"compiler" : 
-			{
-				"id" : "AppleClang",
-				"implicit" : 
-				{
-					"includeDirectories" : 
-					[
-						"/Library/Developer/CommandLineTools/usr/lib/clang/15.0.0/include",
-						"/Library/Developer/CommandLineTools/SDKs/MacOSX14.4.sdk/usr/include",
-						"/Library/Developer/CommandLineTools/usr/include"
-					],
-					"linkDirectories" : 
-					[
-						"/Library/Developer/CommandLineTools/SDKs/MacOSX14.4.sdk/usr/lib",
-						"/Library/Developer/CommandLineTools/SDKs/MacOSX14.4.sdk/usr/lib/swift"
-					],
-					"linkFrameworkDirectories" : 
-					[
-						"/Library/Developer/CommandLineTools/SDKs/MacOSX14.4.sdk/System/Library/Frameworks"
-					],
-					"linkLibraries" : []
-				},
-				"path" : "/Library/Developer/CommandLineTools/usr/bin/cc",
-				"version" : "15.0.0.15000309"
-			},
-			"language" : "C",
-			"sourceFileExtensions" : 
-			[
-				"c",
-				"m"
-			]
-		}
-	],
-	"version" : 
-	{
-		"major" : 1,
-		"minor" : 0
-	}
-}
