+++ cmake-build-debug/CMakeFiles/clion-Debug-log.txt
@@ -1,10 +1,4 @@
 /Applications/CLion.app/Contents/bin/cmake/mac/aarch64/bin/cmake -DCMAKE_BUILD_TYPE=Debug -DCMAKE_MAKE_PROGRAM=/Applications/CLion.app/Contents/bin/ninja/mac/aarch64/ninja -G Ninja -S "/Users/owner/Storage/KSE/IT Disciplines/PP/simple-text-editor" -B "/Users/owner/Storage/KSE/IT Disciplines/PP/simple-text-editor/cmake-build-debug"
--- The C compiler identification is AppleClang 15.0.0.15000309
--- Detecting C compiler ABI info
--- Detecting C compiler ABI info - done
--- Check for working C compiler: /Library/Developer/CommandLineTools/usr/bin/cc - skipped
--- Detecting C compile features
--- Detecting C compile features - done
--- Configuring done (0.4s)
+-- Configuring done (0.0s)
 -- Generating done (0.0s)
 -- Build files have been written to: /Users/owner/Storage/KSE/IT Disciplines/PP/simple-text-editor/cmake-build-debug
