+++ cmake-build-debug/Testing/Temporary/LastTest.log
@@ -1,3 +1,3 @@
-Start testing: May 12 18:10 EEST
+Start testing: May 16 23:41 EEST
 ----------------------------------------------------------
-End testing: May 12 18:10 EEST
+End testing: May 16 23:41 EEST
