+++ include/parser.h
@@ -1,0 +1,9 @@
+#ifndef SIMPLE_TEXT_EDITOR_PARSER_H
+#define SIMPLE_TEXT_EDITOR_PARSER_H
+
+// functions declarations
+void printCommandsInfo();
+void getValidCommand(int* command);
+void executeCommand(int command);
+
+#endif // SIMPLE_TEXT_EDITOR_PARSER_H
